+++ packages/app/src/specs/CreateSpecCard.vue
@@ -5,9 +5,10 @@
     :description="description"
     :icon="icon"
     :icon-size="48"
-    class="w-280px m-2px min-h-216px max-h-350px px-32px pt-32px
-  pb-24px"
+    class="m-2px min-h-216px max-h-350px px-32px pt-32px pb-24px
+  w-280px"
     variant="indigo"
+    :badge-text="badgeText"
     @click="emits('click')"
   />
 </template>
@@ -21,6 +22,7 @@ defineProps<{
   header: string
   description: string
   disabled?: boolean
+  badgeText?: string
 }>()
 
 const emits = defineEmits<{