+++ packages/data-context/src/actions/LocalSettingsActions.ts
@@ -13,7 +13,7 @@ export interface LocalSettingsApiShape {
 export class LocalSettingsActions {
   constructor (private ctx: DataContext) {}
 
-  setPreferences (stringifiedJson: string) {
+  setPreferences (stringifiedJson: string, type: 'global' | 'project') {
     const toJson = JSON.parse(stringifiedJson) as AllowedState
 
     // update local data
@@ -21,8 +21,13 @@ export class LocalSettingsActions {
       this.ctx.coreData.localSettings.preferences[key as keyof AllowedState] = value as any
     }
 
-    // persist to appData
-    return this.ctx._apis.localSettingsApi.setPreferences(toJson)
+    if (type === 'global') {
+      // persist to global appData - projects/__global__/state.json
+      return this.ctx._apis.localSettingsApi.setPreferences(toJson)
+    }
+
+    // persist to project appData - for example projects/launchpad/state.json
+    return this.ctx._apis.projectApi.setProjectPreferences(toJson)
   }
 
   async refreshLocalSettings () {