+++ packages/driver/cypress/e2e/cypress/cy.cy.js
@@ -113,7 +113,8 @@ describe('driver/src/cypress/cy', () => {
       })
     })
 
-    it('stores invocation stack for first command', () => {
+    // TODO(webkit): fix+unskip for experimental webkit
+    it('stores invocation stack for first command', { browser: '!webkit' }, () => {
       cy
       .get('input:first')
       .then(() => {
@@ -123,7 +124,8 @@ describe('driver/src/cypress/cy', () => {
       })
     })
 
-    it('stores invocation stack for chained command', () => {
+    // TODO(webkit): fix+unskip for experimental webkit
+    it('stores invocation stack for chained command', { browser: '!webkit' }, () => {
       cy
       .get('div')
       .find('input')
@@ -173,7 +175,8 @@ describe('driver/src/cypress/cy', () => {
       })
     })
 
-    describe('invocation stack', () => {
+    // TODO(webkit): fix+unskip for experimental webkit
+    describe('invocation stack', { browser: '!webkit' }, () => {
       beforeEach(() => {
         Cypress.Commands.add('getInput', () => cy.get('input'))
         Cypress.Commands.add('findInput', { prevSubject: 'element' }, (subject) => {