+++ packages/launchpad/cypress/e2e/choose-a-browser.cy.ts
@@ -1,6 +1,7 @@
 import type { FoundBrowser } from '@packages/types/src'
 
-describe('Choose a Browser Page', () => {
+// TODO: fix flaky tests https://github.com/cypress-io/cypress/issues/23418
+describe.skip('Choose a Browser Page', () => {
   beforeEach(() => {
     cy.scaffoldProject('launchpad')
   })
@@ -14,7 +15,8 @@ describe('Choose a Browser Page', () => {
       })
     })
 
-    it('preselects browser that is provided through the command line', () => {
+    // TODO: fix flaky test https://github.com/cypress-io/cypress/issues/23158
+    it.skip('preselects browser that is provided through the command line', () => {
       cy.withCtx((ctx, o) => {
         // stub launching project since we have `--browser --testingType --project` here
         o.sinon.stub(ctx._apis.projectApi, 'launchProject').resolves()
@@ -120,7 +122,7 @@ describe('Choose a Browser Page', () => {
         body: {
           data: {
             launchOpenProject: true,
-            setProjectPreferences: {
+            setProjectPreferencesInGlobalCache: {
               currentProject: {
                 id: 'test-id',
                 title: 'launchpad',
@@ -215,7 +217,8 @@ describe('Choose a Browser Page', () => {
       cy.wait('@closeBrowser')
     })
 
-    it('performs mutation to focus open browser when focus button is pressed', () => {
+    // TODO: fix flaky test https://github.com/cypress-io/cypress/issues/23220
+    it.skip('performs mutation to focus open browser when focus button is pressed', () => {
       cy.openProject('launchpad', ['--e2e'])
 
       cy.visitLaunchpad()
@@ -267,7 +270,8 @@ describe('Choose a Browser Page', () => {
       })
     })
 
-    it('subscribes to changes to browserStatus/activeBrowser through the browserStatusUpdated subscription', () => {
+    // TODO: fix flaky test https://github.com/cypress-io/cypress/issues/23158
+    it.skip('subscribes to changes to browserStatus/activeBrowser through the browserStatusUpdated subscription', () => {
       cy.openProject('launchpad', ['--e2e'])
 
       cy.visitLaunchpad()