+++ packages/types/src/preferences.ts
@@ -1,4 +1,4 @@
-import type { Editor } from '.'
+import type { BannersState, Editor } from '.'
 
 export const defaultPreferences: AllowedState = {
   autoScrollingEnabled: true,
@@ -12,6 +12,7 @@ export const allowedKeys: Readonly<Array<keyof AllowedState>> = [
   'appX',
   'appY',
   'autoScrollingEnabled',
+  'banners',
   'browserWidth',
   'browserHeight',
   'browserX',
@@ -32,6 +33,7 @@ export const allowedKeys: Readonly<Array<keyof AllowedState>> = [
   'lastOpened',
   'lastProjectId',
   'promptsShown',
+  'specFilter',
   'preferredEditorBinary',
   'isSideNavigationOpen',
   'lastBrowser',
@@ -46,6 +48,7 @@ export type AllowedState = Partial<{
   appY: Maybe<number>
   isSpecsListOpen: Maybe<boolean>
   autoScrollingEnabled: Maybe<boolean>
+  banners: Maybe<BannersState>
   browserWidth: Maybe<number>
   browserHeight: Maybe<number>
   browserX: Maybe<number>
@@ -65,6 +68,7 @@ export type AllowedState = Partial<{
   firstOpened: Maybe<number>
   lastOpened: Maybe<number>
   promptsShown: Maybe<object>
+  specFilter: Maybe<string>
   preferredEditorBinary: Maybe<string>
   isSideNavigationOpen: Maybe<boolean>
   testingType: 'e2e' | 'component'